+++ framework/model/src/main/java/is/codion/framework/model/EntityTableConditionModel.java
@@ -53,7 +53,7 @@ public interface EntityTableConditionModel extends TableConditionModel<Attribute
 	 * Sets the EQUAL condition value of the condition model associated with {@code attribute}.
 	 * Enables the condition model in case {@code value} is non-empty or disables it if {@code values is empty}.
 	 * @param attribute the attribute
-	 * @param values the search condition value
+	 * @param value the search condition value
 	 * @param <T> the value type
 	 * @return true if the search state changed as a result of this method call, false otherwise
 	 */