+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/Dialogs.java
@@ -154,11 +154,10 @@ public final class Dialogs {
 	}
 
 	/**
-	 * @param userPreferencePropertyName the name of the property to use when saving the selected font size as a user preference
 	 * @return a builder for a dialog for selecting the font size
 	 */
-	public static FontSizeSelectionDialogBuilder fontSizeSelectionDialog(String userPreferencePropertyName) {
-		return new DefaultFontSizeSelectionDialogBuilder(userPreferencePropertyName);
+	public static FontSizeSelectionDialogBuilder fontSizeSelectionDialog() {
+		return new DefaultFontSizeSelectionDialogBuilder();
 	}
 
 	/**