+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/FontSizeSelectionDialogBuilder.java
@@ -21,7 +21,7 @@ package is.codion.swing.common.ui.dialog;
 import is.codion.swing.common.ui.control.Control;
 
 import javax.swing.JComponent;
-import java.util.OptionalInt;
+import java.util.function.Consumer;
 
 /**
  * Builds a dialog for selecting the font size.
@@ -35,14 +35,21 @@ public interface FontSizeSelectionDialogBuilder {
 	FontSizeSelectionDialogBuilder owner(JComponent owner);
 
 	/**
-	 * Displays a dialog allowing the user the select a font size multiplier.
-	 * @return the selected font size multiplier, an empty Optional if cancelled
+	 * @param initialSelection the initally selected font size ratio, default 100%
+	 * @return this builder
 	 */
-	OptionalInt selectFontSize();
+	FontSizeSelectionDialogBuilder initialSelection(int initialSelection);
+
+	/**
+	 * Displays a dialog allowing the user the select a font size multiplier.
+	 * @param selectedFontSize called when the OK button is pressed
+	 */
+	void selectFontSize(Consumer<Integer> selectedFontSize);
 
 	/**
 	 * Creates a {@link Control} for selecting the font size.
+	 * @param selectedFontSize called when the OK button is pressed
 	 * @return a Control for displaying a dialog for selecting a font size
 	 */
-	Control createControl();
+	Control createControl(Consumer<Integer> selectedFontSize);
 }