+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/LookAndFeelSelectionDialogBuilder.java
@@ -22,7 +22,7 @@ import is.codion.swing.common.ui.control.Control;
 import is.codion.swing.common.ui.laf.LookAndFeelProvider;
 
 import javax.swing.JComponent;
-import java.util.Optional;
+import java.util.function.Consumer;
 
 /**
  * Builds a dialog for selecting a look and feel.
@@ -42,20 +42,15 @@ public interface LookAndFeelSelectionDialogBuilder {
 	LookAndFeelSelectionDialogBuilder enableOnSelection(boolean enableOnSelection);
 
 	/**
-	 * @param userPreferencePropertyName the name of the property to use when saving the selected look and feel as a user preference
-	 * @return this builder
-	 */
-	LookAndFeelSelectionDialogBuilder userPreferencePropertyName(String userPreferencePropertyName);
-
-	/**
 	 * Displays a dialog allowing the user the select between all available Look and Feels.
-	 * @return the selected look and feel provider, an empty Optional if cancelled
+	 * @param selectedLookAndFeel called when the OK button is pressed
 	 */
-	Optional<LookAndFeelProvider> selectLookAndFeel();
+	void selectLookAndFeel(Consumer<LookAndFeelProvider> selectedLookAndFeel);
 
 	/**
 	 * Creates a {@link Control} for selecting the Look and Feel.
+	 * @param selectedLookAndFeel called when the OK button is pressed
 	 * @return a Control for displaying a dialog for selecting a look and feel
 	 */
-	Control createControl();
+	Control createControl(Consumer<LookAndFeelProvider> selectedLookAndFeel);
 }