+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -22,6 +22,7 @@
  * {@link is.codion.swing.framework.ui.EntityApplicationPanel#CONFIRM_EXIT}<br>
  * {@link is.codion.swing.framework.ui.EntityApplicationPanel#SHOW_STARTUP_DIALOG}<br>
  * {@link is.codion.swing.framework.ui.EntityApplicationPanel#CACHE_ENTITY_PANELS}<br>
+ * {@link is.codion.swing.framework.ui.EntityApplicationPanel#USE_CLIENT_PREFERENCES}<br>
  * {@link is.codion.swing.framework.ui.EntityEditComponentPanel#MODIFIED_INDICATOR}<br>
  * {@link is.codion.swing.framework.ui.EntityEditComponentPanel#MODIFIED_INDICATOR_UNDERLINE_STYLE}<br>
  * {@link is.codion.swing.framework.ui.EntityEditComponentPanel#DEFAULT_TEXT_FIELD_COLUMNS}<br>
@@ -35,7 +36,6 @@
  * {@link is.codion.swing.framework.ui.EntityPanel.Config#CONTROL_PANEL_CONSTRAINTS}<br>
  * {@link is.codion.swing.framework.ui.EntityPanel.Config#CONTROL_TOOLBAR_CONSTRAINTS}<br>
  * {@link is.codion.swing.framework.ui.EntityPanel.Config#WINDOW_TYPE}<br>
- * {@link is.codion.swing.framework.ui.EntityPanel.Config#USE_CLIENT_PREFERENCES}<br>
  * {@link is.codion.swing.framework.ui.EntityTablePanel.Config#QUERY_HIDDEN_COLUMNS}<br>
  * {@link is.codion.swing.framework.ui.EntityTablePanel.Config#ORDER_QUERY_BY_SORT_ORDER}<br>
  * {@link is.codion.swing.framework.ui.EntityTablePanel.Config#CONDITION_STATE}<br>